Renamed SceneManager files for consistency.
Renamed ht_scenemanager.h/cpp to ht_scenemanager_singleton.h/cpp for
consistency with other singletons.
